requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
openai>=1.0.0
pypdf>=4.0.0
docx2txt>=0.8
numpy>=1.26.0
rank-bm25>=0.2.2
